# swagger.yaml
openapi: 3.0.3
info:
  title: QueryTube API
  description: |
    AI-powered video Q&A system with semantic search capabilities.
    Transform any video into an intelligent chatbot.
  version: 1.0.0
  contact:
    name: Vivek Kumar Singh
    url: https://github.com/reachvivek
  license:
    name: Custom Proprietary License
    url: https://github.com/reachvivek/QueryTube/blob/main/LICENSE.md
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://querytube.app/api
    description: Production server
...
paths:
  /validate-video:
    get:
      tags:
        - Video
      summary: Validate YouTube video URL
      description: Validates a YouTube URL and returns video metadata
      parameters:
        - name: url
          in: query
          required: true
          schema:
            type: string
            example: https://www.youtube.com/watch?v=dQw4w9WgXcQ
          description: YouTube video URL
      responses:
        '200':
          description: Video validated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid:
                    type: boolean
                    example: true
                  video:
                    $ref: '#/components/schemas/VideoMetadata'
        '400':
          description: Invalid URL format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /videos:
    post:
      tags:
        - Video
      summary: Create new video record
      description: Creates a new video entry in the database
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateVideoRequest'
      responses:
        '200':
          description: Video created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  video:
                    $ref: '#/components/schemas/Video'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /videos/{id}:
    get:
      tags:
        - Video
      summary: Get video by ID
      description: Retrieves video details with chunks and analytics
      parameters:
        - $ref: '#/components/parameters/VideoId'
      responses:
        '200':
          description: Video retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  video:
                    $ref: '#/components/schemas/VideoWithRelations'
        '404':
          description: Video not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      tags:
        - Video
      summary: Update video
      description: Updates video metadata, transcript, and chunks
      parameters:
        - $ref: '#/components/parameters/VideoId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateVideoRequest'
      responses:
        '200':
          description: Video updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  video:
                    $ref: '#/components/schemas/Video'

    delete:
      tags:
        - Video
      summary: Delete video
      description: Deletes video, chunks (DB + Pinecone), and analytics
      parameters:
        - $ref: '#/components/parameters/VideoId'
      responses:
        '200':
          description: Video deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Video deleted successfully
                  vectorsDeleted:
                    type: integer
                    example: 30
...
components:
  parameters:
    VideoId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Video UUID
  schemas:
    VideoMetadata:
      type: object
      properties:
        id:
          type: string
          example: dQw4w9WgXcQ
        title:
          type: string
          example: "Rick Astley - Never Gonna Give You Up"
        duration:
          type: integer
          example: 213
          description: Duration in seconds
        durationFormatted:
          type: string
          example: "3:33"
        thumbnail:
          type: string
          format: uri
          example: https://i.ytimg.com/vi/dQw4w9WgXcQ/maxresdefault.jpg
        uploader:
          type: string
          example: Rick Astley
        description:
          type: string
          example: The official video for Never Gonna Give You Up
        url:
          type: string
          format: uri
          example: https://www.youtube.com/watch?v=dQw4w9WgXcQ
    Video:
      type: object
      properties:
        id:
          type: string
          format: uuid
        youtubeUrl:
          type: string
          format: uri
        youtubeId:
          type: string
        title:
          type: string
        description:
          type: string
        duration:
          type: integer
        durationFormatted:
          type: string
        thumbnail:
          type: string
          format: uri
        uploader:
          type: string
        language:
          type: string
          example: en
        status:
          type: string
          enum:
            - pending
            - processing
            - completed
            - failed
        errorMessage:
          type: string
          nullable: true
        transcriptSource:
          type: string
          enum:
            - youtube-captions
            - whisper-api
            - uploaded-file
          nullable: true
        transcript:
          type: string
          nullable: true
        summary:
          type: string
          nullable: true
        uploadedAt:
          type: string
          format: date-time
        processedAt:
          type: string
          format: date-time
          nullable: true
    VideoWithRelations:
      allOf:
        - $ref: '#/components/schemas/Video'
        - type: object
          properties:
            chunks:
              type: array
              items:
                $ref: '#/components/schemas/ChunkWithVectorId'
            analytics:
              type: array
              items:
                $ref: '#/components/schemas/Analytics'
    Chunk:
      type: object
      properties:
        chunkIndex:
          type: integer
          example: 0
        text:
          type: string
          example: In this video, we'll explore quantum mechanics...
        startTime:
          type: integer
          example: 0
          description: Start time in seconds
        endTime:
          type: integer
          example: 90
          description: End time in seconds
    ChunkWithVectorId:
      allOf:
        - $ref: '#/components/schemas/Chunk'
        - type: object
          properties:
            id:
              type: string
              format: uuid
            vectorId:
              type: string
              nullable: true
            embeddingProvider:
              type: string
              nullable: true
              example: mistral
            timestamp:
              type: string
              example: "0:00"
            createdAt:
              type: string
              format: date-time
    Analytics:
      type: object
      properties:
        id:
          type: string
          format: uuid
        videoId:
          type: string
          format: uuid
        question:
          type: string
        answer:
          type: string
        responseTime:
          type: number
          format: float
          example: 1.2
        provider:
          type: string
          example: groq
          enum:
            - groq
            - openai
            - claude
        model:
          type: string
          example: llama-3.3-70b-versatile
        chunksUsed:
          type: integer
          example: 5
        timestamp:
          type: string
          format: date-time
    CreateVideoRequest:
      type: object
      required:
        - youtubeUrl
        - title
        - duration
        - durationFormatted
      properties:
        youtubeUrl:
          type: string
          format: uri
        youtubeId:
          type: string
        title:
          type: string
        description:
          type: string
        duration:
          type: integer
        durationFormatted:
          type: string
        thumbnail:
          type: string
          format: uri
        uploader:
          type: string
    UpdateVideoRequest:
      type: object
      properties:
        transcript:
          type: string
        transcriptSource:
          type: string
        language:
          type: string
        status:
          type: string
          enum:
            - pending
            - processing
            - completed
            - failed
        processedAt:
          type: string
          format: date-time
        chunks:
          type: array
          items:
            $ref: '#/components/schemas/Chunk'
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: Invalid request
        message:
          type: string
          example: Detailed error message